Added sample unit testing file, refactored existing NameNodeTest code to support this file separation
