+++ test/CMakeLists.txt
@@ -27,7 +27,7 @@ link_directories(/home/vagrant/rdfs/build/zkwrapper)
 
 add_executable(RunAllTests run-all/run-all-tests.cc)
 
-add_executable(NameNodeTest namenode/NameNodeTest.cc)
+add_executable(NameNodeTest namenode/NameNodeTest.cc namenode/ListingTest.cc namenode/NameNodeTest.h)
 target_link_libraries(NameNodeTest namenode_protos)
 target_link_libraries(NameNodeTest zk_client_nn zk_client_dn)
 target_link_libraries(NameNodeTest zkwrapper /usr/local/lib/libzookeeper_mt.so)
