+++ test/namenode/ListingTest.cc
@@ -1,0 +1,7 @@
+// Copyright 2017 Rice University, COMP 413 2017
+
+#include "NameNodeTest.h"
+
+TEST_F(NamenodeTest, getOneFileListing) {
+
+}
